requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
websocket-client>=1.6.0
aiohttp>=3.9.0
asyncio>=3.4.3
logging>=0.4.9.6
